remove VersionInWorkspace in favor of Exported absoluteProjectRoot and new gps.ProjectVersion
