+++ context_test.go
@@ -14,7 +14,6 @@ import (
 	"testing"
 	"unicode"
 
-	"github.com/golang/dep/internal/gps"
 	"github.com/golang/dep/internal/test"
 )
 
@@ -80,7 +79,7 @@ func TestAbsoluteProjectRoot(t *testing.T) {
 	}
 
 	for i, ok := range importPaths {
-		got, err := depCtx.absoluteProjectRoot(i)
+		got, err := depCtx.AbsForImport(i)
 		if ok {
 			h.Must(err)
 			want := h.Path(filepath.Join("src", i))
@@ -97,54 +96,9 @@ func TestAbsoluteProjectRoot(t *testing.T) {
 
 	// test that a file fails
 	h.TempFile("src/thing/thing.go", "hello world")
-	_, err := depCtx.absoluteProjectRoot("thing/thing.go")
+	_, err := depCtx.AbsForImport("thing/thing.go")
 	if err == nil {
 		t.Fatal("error should not be nil for a file found")
-	}
-}
-
-func TestVersionInWorkspace(t *testing.T) {
-	test.NeedsExternalNetwork(t)
-	test.NeedsGit(t)
-
-	h := test.NewHelper(t)
-	defer h.Cleanup()
-
-	h.TempDir("src")
-	h.Setenv("GOPATH", h.Path("."))
-	depCtx := &Ctx{GOPATH: h.Path(".")}
-
-	importPaths := map[string]struct {
-		rev      gps.Version
-		checkout bool
-	}{
-		"github.com/pkg/errors": {
-			rev:      gps.NewVersion("v0.8.0").Pair("645ef00459ed84a119197bfb8d8205042c6df63d"), // semver
-			checkout: true,
-		},
-		"github.com/Sirupsen/logrus": {
-			rev:      gps.Revision("42b84f9ec624953ecbf81a94feccb3f5935c5edf"), // random sha
-			checkout: true,
-		},
-		"github.com/rsc/go-get-default-branch": {
-			rev: gps.NewBranch("another-branch").Pair("8e6902fdd0361e8fa30226b350e62973e3625ed5"),
-		},
-	}
-
-	// checkout the specified revisions
-	for ip, info := range importPaths {
-		h.RunGo("get", ip)
-		repoDir := h.Path("src/" + ip)
-		if info.checkout {
-			h.RunGit(repoDir, "checkout", info.rev.String())
-		}
-
-		got, err := depCtx.VersionInWorkspace(gps.ProjectRoot(ip))
-		h.Must(err)
-
-		if got != info.rev {
-			t.Fatalf("expected %q, got %q", got.String(), info.rev.String())
-		}
 	}
 }
